% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exVCF}
\alias{exVCF}
\title{VCF file from chromosome 20}
\format{a CollapsedVCF with 43 variants and annotations}
\usage{
exVCF
}
\description{
A dataset in VCF format containing variants called with HaplotypeCaller
Only on chromosome 20
}
\keyword{datasets}
