% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{vcf_histogram}
\alias{vcf_histogram}
\title{Plotting Histogram of Variant consequences}
\usage{
vcf_histogram(vcf, ...)
}
\arguments{
\item{vcf}{an object of class VcfFile}

\item{...}{additional parameters for the plotting}
}
\value{
A \code{\link{ggplot2}} plot object
}
\description{
Plotting Histogram of Variant consequences
}
\examples{
\dontrun{
vcf_histogram(vcf)
}
}
